+++ Runtime/Subsystems/Camera/ARCameraFeature.cs
@@ -70,7 +70,7 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
         }
 
         /// <summary>
-        /// Instantiates Meta OpenXR Camera subsystem instances, but does not start it.
+        /// Creates the <see cref="MetaOpenXRCameraSubsystem"/>, but does not start it.
         /// (Start/Stop is typically handled by AR Foundation managers.)
         /// </summary>
         protected override void OnSubsystemCreate()