+++ Runtime/Subsystems/Mesh/ARMeshFeature.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using UnityEditor;
 using UnityEditor.XR.OpenXR.Features;
 #endif
+using static UnityEngine.XR.OpenXR.Features.Meta.Constants.OpenXRExtensions;
+using static UnityEngine.XR.OpenXR.Features.Meta.SystemCapabilityUtils;
 
 namespace UnityEngine.XR.OpenXR.Features.Meta
 {
@@ -10,7 +12,7 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
     /// Enables AR Foundation meshing support via OpenXR for Meta Quest devices.
     /// </summary>
 #if UNITY_EDITOR
-    [OpenXRFeature(UiName = "Meta Quest: Meshing",
+    [OpenXRFeature(UiName = k_DisplayName,
         BuildTargetGroups = new[] { BuildTargetGroup.Android, BuildTargetGroup.Standalone },
         Company = Constants.k_CompanyName,
         Desc = "AR Foundation mesh support on Meta Quest devices",
@@ -22,12 +24,13 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
 #endif
     public class ARMeshFeature : MetaOpenXRFeature
     {
+        const string k_DisplayName = "Meta Quest: Meshing";
+        const string k_SubsystemId = "MetaOpenXR-Meshing";
+
         /// <summary>
         /// The feature id string. This is used to give the feature a well known id for reference.
         /// </summary>
         public const string featureId = "com.unity.openxr.feature.arfoundation-meta-mesh";
-
-        const string k_SubsystemId = "MetaOpenXR-Meshing";
 
         /// <summary>
         /// The set of OpenXR spec extension strings to enable, separated by spaces.
@@ -36,15 +39,40 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
         /// </summary>
         /// <seeaslo href="https://developer.oculus.com/documentation/native/android/mobile-scene-api-ref/"/>
         const string k_OpenXRRequestedExtensions =
-            Constants.OpenXRExtensions.k_XR_FB_spatial_entity + " " +
-            Constants.OpenXRExtensions.k_XR_FB_scene + " " +
-            Constants.OpenXRExtensions.k_XR_META_spatial_entity_mesh + " " +
-            Constants.OpenXRExtensions.k_XR_META_spatial_entity_discovery;
+            k_XR_FB_scene + " " +
+            k_XR_FB_spatial_entity + " " +
+            k_XR_META_spatial_entity_discovery + " " +
+            k_XR_META_spatial_entity_mesh;
 
         static List<XRMeshSubsystemDescriptor> s_MeshDescriptors = new();
 
         /// <summary>
-        /// Instantiates Meta OpenXR mesh subsystem instance, but does not start it.
+        /// Called after `xrCreateInstance`. Override this method to validate that any necessary OpenXR extensions were
+        /// successfully enabled
+        /// (<a href="xref:UnityEngine.XR.OpenXR.OpenXRRuntime.IsExtensionEnabled(System.String)">OpenXRRuntime.IsExtensionEnabled</a>)
+        /// and that any required system properties are supported. If this method returns <see langword="false"/>,
+        /// the feature's [enabled](xref:UnityEngine.XR.OpenXR.Features.OpenXRFeature.enabled) property is set to <see langword="false"/>.
+        /// </summary>
+        /// <param name="xrInstance">Handle of the native `xrInstance`.</param>
+        /// <returns><see langword="true"/> if this feature successfully initialized. Otherwise, <see langword="false"/>.</returns>
+        /// <remarks>
+        /// If this feature is a required feature of an enabled feature set, returning <see langword="false"/> here
+        /// causes the `OpenXRLoader` to fail, and XR Plug-in Management will fall back to another loader if enabled.
+        /// </remarks>
+        /// <seealso href="xref:openxr-features#enabling-openxr-spec-extension-strings">Enabling OpenXR spec extension strings</seealso>
+        protected override bool OnInstanceCreate(ulong xrInstance)
+        {
+            return
+                OpenXRRuntime.IsExtensionEnabled(k_XR_FB_scene)
+                && OpenXRRuntime.IsExtensionEnabled(k_XR_FB_spatial_entity)
+                && OpenXRRuntime.IsExtensionEnabled(k_XR_META_spatial_entity_discovery)
+                && OpenXRRuntime.IsExtensionEnabled(k_XR_META_spatial_entity_mesh)
+                && IsCapabilitySupported(SystemCapability.SpatialEntity, xrInstance, k_DisplayName, typeof(MetaOpenXRBoundingBoxSubsystem))
+                && IsCapabilitySupported(SystemCapability.SpaceDiscovery, xrInstance, k_DisplayName, typeof(MetaOpenXRBoundingBoxSubsystem));
+        }
+
+        /// <summary>
+        /// Instantiates the meshing subsystem instance, but does not start it.
         /// (Start/Stop is typically handled by AR Foundation managers.)
         /// </summary>
         protected override void OnSubsystemCreate()